refactor: move expenses from apps to components
