+++ src/components/expenses/ExpensesWithData.js
@@ -1,11 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { graphql } from 'react-apollo';
 import gql from 'graphql-tag';
+import { graphql } from 'react-apollo';
 import { get } from 'lodash';
 
-import withIntl from '../../../lib/withIntl';
-import Error from '../../../components/Error';
+import withIntl from '../../lib/withIntl';
+
+import Error from '../Error';
 import Expenses from './Expenses';
 
 class ExpensesWithData extends React.Component {
