+++ src/components/expenses/Orders.js
@@ -2,7 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { ButtonGroup, Button } from 'react-bootstrap';
 import { FormattedMessage } from 'react-intl';
-import colors from '../../../constants/colors';
+
+import colors from '../../constants/colors';
 
 import Order from './Order';
