+++ src/components/expenses/RefundTransactionBtn.js
@@ -4,10 +4,10 @@ import gql from 'graphql-tag';
 import { graphql } from 'react-apollo';
 import { FormattedMessage } from 'react-intl';
 
-import withIntl from '../../../lib/withIntl';
-import { getTransactionsQuery, transactionFields } from '../../../graphql/queries';
+import withIntl from '../../lib/withIntl';
+import { getTransactionsQuery, transactionFields } from '../../graphql/queries';
 
-import SmallButton from '../../../components/SmallButton';
+import SmallButton from '../SmallButton';
 
 class RefundTransactionBtn extends React.Component {
   static propTypes = {
