+++ src/components/expenses/Transactions.js
@@ -2,15 +2,14 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { ButtonGroup, Button } from 'react-bootstrap';
 import { FormattedMessage } from 'react-intl';
+import { Box } from '@rebass/grid';
 
-import withIntl from '../../../lib/withIntl';
-import colors from '../../../constants/colors';
+import withIntl from '../../lib/withIntl';
+import colors from '../../constants/colors';
 
 import Transaction from './Transaction';
 import TransactionsExportPopoverAndButton from './TransactionsExportPopoverAndButton';
 import DownloadInvoicesPopOver from './DownloadInvoicesPopOver';
-
-import { Box } from '@rebass/grid';
 
 class Transactions extends React.Component {
   static propTypes = {
