+++ src/pages/createExpense.js
@@ -6,8 +6,8 @@ import gql from 'graphql-tag';
 import { Box, Flex } from '@rebass/grid';
 import { pick } from 'lodash';
 
-import ExpensesStatsWithData from '../apps/expenses/components/ExpensesStatsWithData';
-import CreateExpenseForm from '../apps/expenses/components/CreateExpenseForm';
+import ExpensesStatsWithData from '../components/expenses/ExpensesStatsWithData';
+import CreateExpenseForm from '../components/expenses/CreateExpenseForm';
 
 import ErrorPage from '../components/ErrorPage';
 import Button from '../components/Button';
